% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackSampling.R
\name{flip01}
\alias{flip01}
\title{Flip a 1 and a 0 simultaneously}
\usage{
flip01(x)
}
\arguments{
\item{x}{an integer or logical vector}
}
\value{
x an integer vector
}
\description{
Flip a 1 and a 0 simultaneously
}

