% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackSampling.R
\name{\%[[\%}
\alias{\%[[\%}
\title{Accessing the same named list elements of a list of lists}
\usage{
x \%[[\% n
}
\arguments{
\item{x}{- a list of sublists with same elements in each sublist}

\item{n}{- name of element to be accessed from each sublist}
}
\value{
list of elements of name n in each sublist
}
\description{
Accessing the same named list elements of a list of lists
}
\references{
credits to @kohske, Accessing same named list elements of the list of lists in R,
http://stackoverflow.com/questions/5935673/accessing-same-named-list-elements-of-the-list-of-lists-in-r/5936077#5936077
}

