% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackSampling.R
\name{initState}
\alias{initState}
\title{Generate an initial feasible solution by solving a linear programming with binary variables}
\usage{
initState(numVar, objVec = runif(numVar), constraints = NULL)
}
\arguments{
\item{numVar}{- number of variables}

\item{objVec}{- objective function as a numeric vector}

\item{constraints}{- a list of list of constraints with constr.mat, constr.dir, constr.rhs in each sublist}
}
\value{
a binary vector containing a feasible solution
}
\description{
Generate an initial feasible solution by solving a linear programming with binary variables
}
\examples{
#see documentation for sampl.mcmc

}

