% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsackSampling.R
\name{unlist.constr}
\alias{unlist.constr}
\title{Unpack constraints in a list of list into a matrix, equality or inequality signs, constant on right hand side}
\usage{
unlist.constr(constraints)
}
\arguments{
\item{constraints}{- a list of list of constraints with constr.mat, constr.dir, constr.rhs in each sublist}
}
\value{
a list containing matrix, signs and RHS constants
}
\description{
Unpack constraints in a list of list into a matrix, equality or inequality signs, constant on right hand side
}

